schoolproject/lgms/requirements.txt
Babel==2.6.0
Django==2.0.6
django-crispy-forms==1.7.2
django-phonenumber-field==2.0.0
phonenumberslite==8.9.8
pytz==2018.4
